+++ src/main/java/app/workout/AddWorkoutDto.java
@@ -2,9 +2,11 @@ package app.workout;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
  record AddWorkoutDto(
         String activityid,
-        Double activitymin
+        BigDecimal activitymin
 ) {
 }
