+++ src/main/java/app/workout/EditWorkoutDto.java
@@ -2,8 +2,10 @@ package app.workout;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
  record EditWorkoutDto(
         Long id,
-        Double activitymin) {
+        BigDecimal activitymin) {
 }
